+++ hotpants.py
@@ -979,7 +979,7 @@ def check_psf_center_numba(iData, imax, jmax, xLen, yLen, sx0, sy0,
                              hiThresh, sky, invdsky,
                              xbuffer, ybuffer, bbit, bbit1,
                              rPixX, hwKSStamp, mRData, kerFitThresh):
-    kerFitThresh = np.float64(np.float32(kerFitThresh))
+    kerFitThresh = np.float64(kerFitThresh)
     brk = 0
     dmax2 = 0.0
     for l in range(jmax - hwKSStamp, jmax + hwKSStamp + 1):
@@ -2504,8 +2504,8 @@ def build_matrix_jit(all_mat, all_vectors, valid_mask, all_x, all_y,
                      wxy, matrix,
                      nS, nCompKer, kerOrder, bgOrder, fwKSStamp, rPixX, rPixY,
                      ncomp, ncomp1, ncomp2, nbg_vec, mat_size, pixStamp):
-    rPixX2 = np.float64(np.float32(0.5 * rPixX))
-    rPixY2 = np.float64(np.float32(0.5 * rPixY))
+    rPixX2 = np.float64(0.5 * rPixX)
+    rPixY2 = np.float64(0.5 * rPixY)
 
     for istamp in range(nS):
         if valid_mask[istamp] == 0:
@@ -2513,8 +2513,8 @@ def build_matrix_jit(all_mat, all_vectors, valid_mask, all_x, all_y,
 
         xstamp = all_x[istamp]
         ystamp = all_y[istamp]
-        fx = np.float64(np.float32(np.float32(xstamp) - np.float32(rPixX2)) / np.float32(rPixX2))
-        fy = np.float64(np.float32(np.float32(ystamp) - np.float32(rPixY2)) / np.float32(rPixY2))
+        fx = np.float64((np.float64(xstamp) - rPixX2) / rPixX2)
+        fy = np.float64((np.float64(ystamp) - rPixY2) / rPixY2)
 
         kk = 0
         a1 = 1.0
@@ -3053,8 +3053,8 @@ def fill_stamp_numba_kernel(
     out_sum_val[0] = sumVal
 
     # ========== Step 3: background vectors ==========
-    rPixX2 = np.float64(np.float32(0.5 * rPixX))
-    rPixY2 = np.float64(np.float32(0.5 * rPixY))
+    rPixX2 = np.float64(0.5 * rPixX)
+    rPixY2 = np.float64(0.5 * rPixY)
     for y_offset in range(fwKSStamp):
         j = yi - hwKSStamp + y_offset
         yf = (j - rPixY2) / rPixY2
